Create random_walks
A simulation of many random walks. In this example, you are climbing a tall building that is more than 60 steps high, using a die to determine the number of steps you take each instance. [1, 2] takes a step down, [3, 4, 5] takes a step up, and [6] allows you to roll the die again and step up the same value as the die roll. Furthermore, there is a 0.1% chance that you slip and fall down to the bottom step, and you cannot go below step 0. This code simulates this walk 500 times, allows you to plot the walks and its distribution, and calculates the probability that you will end up at step 60 or higher by 100 rolls of the die.