Create get_tweet_data
A follow-up on get_post_info, this code uses the Twitter API to collect streaming tweets (filtered by chosen keywords) and save them to a txt file, and then counts the frequency of those keywords in those tweets. It then plots a histogram of the keywords and their frequency for easy visualization. Good for tracking trending keywords and topics within Twitter.